refactor(organization): remove redundant organization component
