+++ frontend/src/common/interfaces/organization.interface.ts
@@ -1,0 +1,9 @@
+export interface IOrganization {
+  id: string;
+  name: string;
+  email: string;
+  phone: string;
+  address: string;
+  logo: string;
+  description: string;
+}
